Move AlertLevel to contract_utils
