--- src/importlinter/domain/output.py
@@ -1,8 +1,0 @@
-import enum
-
-
-@enum.unique
-class AlertLevel(enum.Enum):
-    NONE = "none"
-    WARN = "warn"
-    ERROR = "error"
